pyproject.toml
[project]
name = "ai-secret-game"
version = "0.1.0"
description = "A game for AI agents to trade secrets with different scoring modes"
authors = []
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }

dependencies = [
    "click>=8.1.7",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.4",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "B", "I"]
ignore = []
extend-exclude = ["tests"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"

[project.scripts]
ai-secret-game = "ai_secret_game.cli:cli" 